+++ src/test/scala/scorex/lagonaki/LagonakiTestSuite.scala
@@ -23,6 +23,7 @@ class LagonakiTestSuite extends Suites(
   , new PeersAPISpecification
   , new WalletAPISpecification
   , new AdressesAPISpecification
+  , new TransactionsAPISpecification
 
 
 ) with BeforeAndAfterAll {